Starting to fill the motor update rates vector from the yaml config file. Doesn't seem like the best solution :S

+++ shadow_robot/sr_hand_palm_edc/sr_edc_ethercat_drivers/include/sr_edc_ethercat_drivers/utils/motor_updater.hpp
@@ -47,7 +47,7 @@ namespace motor_updater
   struct UpdateConfig
   {
     FROM_MOTOR_DATA_TYPE what_to_update;
-    int when_to_update;
+    double when_to_update;
   };
 
   class MotorUpdater
